Rename emailBodies to nofificationBodies
